fix: change auth
