+++ src/api/axios-wrapper.ts
@@ -1,21 +1,16 @@
 import axios from "axios";
+import Cookies from "js-cookie";
 
 const axiosInstance = axios.create({
   baseURL: process.env.API_URL,
 });
 
-axiosInstance.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem("jwtToken");
-    if (token) {
-      config.headers.Authorization = token;
-      console.log(token);
-    }
-    return config;
-  },
-  (error) => {
-    return Promise.reject(error);
-  },
-);
+axiosInstance.interceptors.request.use((config) => {
+  const token = Cookies.get("jwtToken");
+  if (token) {
+    config.headers.Authorization = token;
+  }
+  return config;
+});
 
 export default axiosInstance;
