+++ src/api/hooks/query/useAllLotsQuery.ts
@@ -6,7 +6,7 @@ import { toast } from "sonner";
 import { Lots } from "@/api/services/lots";
 import { AllLots } from "@/api/types/lot";
 
-const useGetAllLots = ({
+const useAllLotsQuery = ({
   onSuccessCallback,
   ...rest
 }: {
@@ -28,4 +28,4 @@ const useGetAllLots = ({
   return [data, isLoading, isError];
 };
 
-export default useGetAllLots;
+export default useAllLotsQuery;
