+++ src/api/types/bid-history.ts
@@ -1,0 +1,7 @@
+export type BidHistory = {
+  id: number;
+  amount: number;
+  username: string;
+  time: Date;
+  userEmail: string;
+};
