+++ src/api/types/registration-credentials.ts
